+++ tests/ConfigTest.php
@@ -1,0 +1,68 @@
+<?php
+
+use Expay\SDK\Utility\Config;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * ConfigTest
+ */
+class ConfigTest extends TestCase
+{
+  /**
+   * config
+   *
+   * @var mixed
+   */
+  private $config;
+
+  /**
+   * __construct
+   *
+   * @return void
+   */
+  public function __construct()
+  {
+    parent::__construct();
+    $this->config = new Config("121", "VYYG9HyJhf4Bfzv6qGUx");
+  }
+    
+  /**
+   * testGetSandboxUrl
+   *
+   * @return void
+   */
+  public function testGetSandboxUrl()
+  {
+    $this->assertIsString($this->config->get_sandbox_url());
+  }
+  
+  /**
+   * testGetProductionUrl
+   *
+   * @return void
+   */
+  public function testGetProductionUrl()
+  {
+    $this->assertIsString($this->config->get_production_url());
+  }
+  
+  /**
+   * testGetMerchantId
+   *
+   * @return void
+   */
+  public function testGetMerchantId()
+  {
+    $this->assertIsString($this->config->get_merchant_id());
+  }
+  
+  /**
+   * testGetMerchantKey
+   *
+   * @return void
+   */
+  public function testGetMerchantKey()
+  {
+    $this->assertIsString($this->config->get_merchant_key());
+  }
+}