+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 ## Changes - 02/07/2020 - Jeffery Osei
 * Init composer package requirements
+
+-----------
+
+## Changes - 05/07/2020 - Jeffery Osei
+* Write utilities, requests and exception methods
+* Write test cases